Move FormUtils to a different place
